+++ tests/test_dataset.py
@@ -1,8 +1,8 @@
 import unittest
+from ..CloudHarvestCoreTasks.dataset import DataSet, perform_maths_operation
 
 class TestDataSet(unittest.TestCase):
     def setUp(self):
-        from ..CloudHarvestCoreTasks.dataset import DataSet
 
         self.dataset = DataSet()
         self.dataset.add_records([
@@ -97,6 +97,42 @@ class TestDataSet(unittest.TestCase):
             self.assertIn('street', record['address'])
             self.assertIn('city', record['address'])
 
+    def test_match_and_remove(self):
+        self.dataset.match_and_remove(['name == "John Doe"'])
+        for record in self.dataset:
+            self.assertNotEqual(record['name'], 'John Doe')
+
+    def test_maths_keys(self):
+        self.dataset.maths_keys(source_keys=['age', 'age'],
+                                target_key='age_plus_age',
+                                operation='add',
+                                missing_value=0,
+                                default_value=None)
+
+        for record in self.dataset:
+            self.assertIn('age_plus_age', record)
+            self.assertEqual(record['age_plus_age'], record['age'] * 2)
+
+    def test_maths_records(self):
+        self.dataset.maths_records(source_key='age',
+                                   target_key='average_age',
+                                   operation='average',
+                                   missing_value=0,
+                                   default_value=None)
+
+        self.assertIn('average_age', self.dataset.maths_results)
+        self.assertEqual(self.dataset.maths_results['average_age'], 32.5)
+
+    def test_append_maths_results(self):
+        self.dataset.maths_records(source_key='age',
+                                   target_key='average_age',
+                                   operation='average',
+                                   missing_value=0,
+                                   default_value=None)
+
+        self.dataset.append_record_maths_results()
+        self.assertEqual(self.dataset[-1]['_id'], 'Totals')
+
     def test_rename_keys(self):
         self.dataset.rename_keys([{'old': 'name', 'new': 'full_name'}])
         for record in self.dataset:
@@ -119,14 +155,8 @@ class TestDataSet(unittest.TestCase):
         for record in self.dataset:
             self.assertIsInstance(record['tags'], list)
 
-    def test_match_and_remove(self):
-        self.dataset.match_and_remove(['name == "John Doe"'])
-        for record in self.dataset:
-            self.assertNotEqual(record['name'], 'John Doe')
-
 class TestWalkableDict(unittest.TestCase):
     def setUp(self):
-        from ..CloudHarvestCoreTasks.dataset import DataSet
 
         self.dataset = DataSet([
             {
@@ -176,6 +206,49 @@ class TestWalkableDict(unittest.TestCase):
         for record in self.dataset:
             record.drop('address.city')
             self.assertNotIn('city', record['address'])
+
+
+class TestPerformMathsOperation(unittest.TestCase):
+
+    def test_add_operation(self):
+        result = perform_maths_operation('add', [1, 2, 3, 4])
+        self.assertEqual(result, 10)
+
+    def test_subtract_operation(self):
+        result = perform_maths_operation('subtract', [10, 2, 3])
+        self.assertEqual(result, -15)
+
+    def test_multiply_operation(self):
+        result = perform_maths_operation('multiply', [2, 3, 4])
+        self.assertEqual(result, 24)
+
+    def test_divide_operation(self):
+        result = perform_maths_operation('divide', [8, 2, 2])
+        self.assertEqual(result, 2)
+
+    def test_average_operation(self):
+        result = perform_maths_operation('average', [2, 4, 6])
+        self.assertEqual(result, 4)
+
+    def test_minimum_operation(self):
+        result = perform_maths_operation('minimum', [2, 4, 6])
+        self.assertEqual(result, 2)
+
+    def test_maximum_operation(self):
+        result = perform_maths_operation('maximum', [2, 4, 6])
+        self.assertEqual(result, 6)
+
+    def test_invalid_operation(self):
+        result = perform_maths_operation('invalid', [2, 4, 6])
+        self.assertIsNone(result)
+
+    def test_missing_values(self):
+        result = perform_maths_operation('add', [])
+        self.assertEqual(result, 0)
+
+    def test_error_data(self):
+        self.assertIsNone(perform_maths_operation('add', [2, 'a', 4]))
+
 
 if __name__ == '__main__':
     unittest.main()